% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discursAr.R
\docType{package}
\name{discurAr}
\alias{discurAr}
\title{\code{polAr} package}
\description{
Caja de Herramientas para el procesamiento de discursos presidenciales de Argentina
See the README on
\href{https://github.com/electorArg/discrusAr/blob/master/README.md}{Github}
}
